Ik ga spagetti eten doei!
Oeps verkeerde ramen.

+++ crafting de pafting poging twee/Assets/Youri Map/Script/AbstactSideEffect.cs
@@ -4,11 +4,7 @@ using UnityEngine;
 
 public abstract class AbstactSideEffect : MonoBehaviour, ISideEffect
 {
-
     public abstract void DrinkEffect();
-
-    public abstract void EffectCount(int addedAmount);
-
-
+    public abstract void EffectCount(float addedAmount);
     public abstract void ThrowEffect();
 }
